Move EndpointBuilder to core project #2
Move EndpointBuilder from BlazorBlog.MicroCms to BlazorBlog.Core because
we want to use it from Strapi project.
