+++ src/BlazorBlog.MicroCms/MicroCmsClient.cs
@@ -2,6 +2,7 @@ using System.Collections.Specialized;
 using System.Net.Http;
 using System.Net.Http.Json;
 using System.Threading.Tasks;
+using BlazorBlog.Helpers;
 
 namespace BlazorBlog.MicroCms
 {
